pyproject.toml
[project]
name = "desmos3"
version = "0.1.0"
description = "Bad Apple tracing utilities"
requires-python = ">=3.10"
dependencies = [
  "flask",
  "flask-cors",
  "numpy",
  "pillow",
]

[tool.uv]
dev-dependencies = []
